Move field parsing out of trigger invocation

+++ test/steps/database.go
@@ -27,7 +27,7 @@ func (s *StepContext) StartDatabase() error {
 	// Create the table we'll be writing to
 	_, err = metricsDB.Exec(`
 		CREATE table user_end (
-			timestamp_ms int,
+			timestamp_ms bigint,
 			connection_token varchar,
 			downloaded_bytes int,
 			uploaded_bytes int,
